String in Table.cpp
