+++ smirnov.alexandr/P2/sinCmath.cpp
